move socket guts into socket class methods
